Workaround for OS X make bug.
OS X's make frequently ends the Rakudo install process with a
"make: write error".  This happens after all the actual work
make is doing is done, so we trap it and issue another
make install (which should fail if there is an actual issue with
the build, and work if everything is already done).