Fix and enhance list-available
First off, the original implementation would not mark the current
version as it was suppose to do:

    $ ./bin/rakudobrew list
    * moar-2014.11
    $ ./bin/rakudobrew list-available | grep '2014\.11'
      2014.11

Second, it does not make much sense, because Rakudo version is a
secondary aspect of a Perl 6 version. Primary one is backend.

So I fixed both, now it lists backends and Rakudo versions, and marks
the current combination with "*"'s like so:

    $ ./bin/rakudobrew current
    Currently running moar-2014.11
    $ ./bin/rakudobrew list-available | grep '\*'
    * moar
    * 2014.11