Allow build-panda to accept a version number
This change makes it possible for a given panda version to be installed
alongside a given moar/nqp/rakudo stack.  For instance, if one wants to
install Rakudo version 2015.04, then it is necessary to use the 2015.04
version of panda (the version is really the tag) since a newer version of
panda would use the Empty keyword which wasn't available in Rakudo 2015.04.
This change makes it possible to debug Perl6 modules across different
versions of Rakudo while also installing the module's dependencies via
panda.