Provide help via POD documentation.
Fix issue #91.

The idea is to let the help branch to introspect the POD documentation
for a specific command in order to see if some documentation can be found.
If documentation can be found, it is shown. This allows for:
- invoking help on subcommands, for instance
  rakudobrew help nuke
- invoking help without any argument, that as already done
  simply prints the list of commands
- invoking an alien command, that results in printing out
  the list of available commands.

This requires an explicit dependency on Pod::Usage.

The help is provided as POD documentation with sections named 'COMMAND'
and the name of the command as passed to rakudobrew. The nuke command
has been introduced as an example.