No need for string concatenation in the run subroutine.
We can interpolate $_[0] into string when croak-ing.
Also provide the exit code of the executed command, this can be
useful for debugging.
Last, since `perldoc -f system` suggests using the 'or' and comparing
the system return value to zero, added a comment to explain that the usage
of 'and' is the same.