Refactor to make handling of different shells easier
The shell hooks are now to be called after the normal rakudobrew
invocation. This has the advantage, that the second shell hook call
can already rely on the result of get_version() and get_brew_mode(),
because the normal invocation has already processed a potential version
or mode change.

General shell related code is now in Rakudobrew/ShellHook.
Code for a specific shell is now in its own module in
Rakudobrew/ShellHook/*.pm.
This gets out a lot more cruft of the main rakudobrew script.

The new code now uses quite a bit of dynamic subroutine calling to
locate and find the correct shell implementation. The only way I was able
to get it working is now using `no strict 'refs';`. Not sure if this can
be done cleaner. But it works.